Refactor: move SprintReader to service.command
