+++ backend/src/main/java/com/tissue/api/sprint/service/command/SprintReader.java
@@ -1,4 +1,4 @@
-package com.tissue.api.sprint.service.query;
+package com.tissue.api.sprint.service.command;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
