+++ backend/src/test/java/com/tissue/support/helper/ServiceIntegrationTestHelper.java
@@ -31,8 +31,8 @@ import com.tissue.api.security.PasswordEncoder;
 import com.tissue.api.sprint.domain.repository.SprintQueryRepository;
 import com.tissue.api.sprint.domain.repository.SprintRepository;
 import com.tissue.api.sprint.service.command.SprintCommandService;
+import com.tissue.api.sprint.service.command.SprintReader;
 import com.tissue.api.sprint.service.query.SprintQueryService;
-import com.tissue.api.sprint.service.query.SprintReader;
 import com.tissue.api.team.domain.repository.TeamRepository;
 import com.tissue.api.team.service.command.TeamCommandService;
 import com.tissue.api.team.service.query.TeamQueryService;
